95 percent there
This is not quite ready to check in but the core logic is 95%.
o delay spawn is fully functional and customizable for allies and enemies.
o delay of enemies is a function of distance from the portal (currently no check for pathing but I think that is a future feature)
o wolves follow after teleportation
o all allies keep their data including name, tamed status, and stars.
TODO
o lots of testing
o replace kludge or else remove name and tame so that they are wild
o add particle effect for delayed spawns